geotrigerring start/stop
